refactor: move UIHandler to View package
